requirements.txt
torchvision
tensorboardX
tqdm
pyaml